LPS-198079 Merge the component DDMPermissionSupportRegistry in DDMPermissionSupportImpl
